requirements.txt
base==0.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.4
pyperclip==1.9.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.97
ultralytics-thop==2.0.6
urllib3==2.2.3
